backend/foodgram/requirements.txt
asgiref==3.5.2
atomicwrites==1.4.0
attrs==19.3.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.0
certifi==2019.9.11
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==4.0.6
django-debug-toolbar==2.2
django-filter==2.4.0
django-templated-mail==1.1.1
django-utils==0.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
Faker==12.0.1
flake8==4.0.1
flake8-isort==4.1.1
gunicorn 
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mixer==7.1.2
more-itertools==8.2.0
mypy==0.931
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==20.1
Pillow==8.3.1
pluggy==0.13.1
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
psycopg2-binary
pyparsing==2.4.6
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2019.3
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.7.0
soupsieve==2.3.2
sqlparse==0.3.0
testfixtures==6.18.5
toml==0.10.2
tomli==2.0.0
typing-extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.25.6
wcwidth==0.1.8
webcolors==1.12
zipp==2.2.0
